Apply OR operator formatting to edit input dialogs
